# swagger-docs.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Postaway Application",
        "title": "E-commerce API"
    },
    "host": "localhost:3000",
    "securityDefinitions": {
        "JWT": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "paths": {
...
        "/api/posts": {
            "post": {
                "tags": ["Post"],
                "summary": "Create a Post",
                "description": "You can create a post",
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "parameters": [
                    {
                        "in": "formData",
                        "name": "imageUrl",
                        "description": "Image file to upload",
                        "type": "file"
                    },
                    {
                        "in": "formData",
                        "name": "caption",
                        "description": "Caption for the post",
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                }
            },
            "get": {
                "tags": ["Post"],
                "summary": "Get All Posts",
                "description": "Retrieve all posts",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    }
}